test: clean up E2E CE/EE split — remove EE-only code from CE files
- mrreview_meta_ce_test.go: fix 3 json struct tags (merge_request_iid → iid)
  to repair mrOut.IID=0 root cause of 3 meta-tool failures.
- runner_meta_ce_test.go: remove runRunnerControllerSubtests (Ultimate-only
  helper + call site + runnercontrollers import) — duplicated in
  runner_meta_ee_test.go which already has inlined ControllerList and
  ControllerGet_Graceful404 subtests. -44 lines.
- groups_meta_*: move the entire group hook E2E subsystem
  (orchestrator + 6 helpers) from groups_meta_helpers_ce_test.go into
  groups_meta_ee_test.go. The CE helpers short-circuited on
  !sess.enterprise so the subsystem was effectively dead code in CE mode
  and never exercised in EE mode either. Add the groups import to the EE
  file, remove the call from groups_meta_ce_test.go, and wire
  runMetaGroupHookOperations into mixed_enterprise_ee_test.go so EE
  sessions actually exercise the hook lifecycle.
- mrapproval_*: extract TestMeta_MRApprovalSettings (Premium-only) into
  a new mrapproval_ee_test.go with the e2e && enterprise build tag. CE
  file keeps TestIndividual_MRApproval and TestMeta_MRApproval. -64 lines
  in CE, +80 lines in new EE file.

Net: -107 LOC. Both build tags (e2e and e2e enterprise) compile
cleanly. CE: 1185 tests, 0 failures, 0 EE-only skips. EE tests now
exercise the moved subsystems instead of silently skipping them.
